+++ lib/http_request_cipher.dart
@@ -1,10 +1,7 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
 library http_request_cipher;
 
-export 'src/http_request_cipher_base.dart';
-export 'src/helper_types.dart';
 export 'src/byte_data_decrypter.dart';
 export 'src/byte_data_encrypter.dart';
+export 'src/helper_types.dart';
+export 'src/http_request_cipher_base.dart';
 export 'src/list_extensions.dart';